+++ iOS Client/DevotionsController/ScheduleDailyDevotionViewController.swift
@@ -7,21 +7,13 @@
 //
 
 import UIKit
-import UserNotifications
-
-enum NotificationPermission {
-    case allowed, denied, error
-}
 
 class ScheduleDailyDevotionViewController: UIViewController {
-    @IBOutlet weak var statusLabel: UILabel!
-    @IBOutlet weak var cancelButton: UIButton!
     @IBOutlet weak var timePicker: UIDatePicker!
-    
     var devotionsStore = MBDevotionsStore()
     var devotions: [LoadedDevotion] = []
-    let center = UNUserNotificationCenter.current()
     let defaultTimeInMinutes = 60 * 8 // 8 a.m.
+    let scheduler: DevotionScheduler = Scheduler()
     
     static func instantiateFromStoryboard() -> ScheduleDailyDevotionViewController {
         // swiftlint:disable force_cast
@@ -34,7 +26,6 @@ class ScheduleDailyDevotionViewController: UIViewController {
         
         // Do any additional setup after loading the view.
         self.title = "Daily Devotional"
-        self.configureBackButton()
         self.devotions = self.devotionsStore.getDevotions()
         
         // configure timePicker
@@ -45,10 +36,6 @@ class ScheduleDailyDevotionViewController: UIViewController {
         var currentSetting: Int = defaultTimeInMinutes
         if let val = UserDefaults.standard.value(forKey: MBConstants.DEFAULTS_DAILY_DEVOTION_TIME_KEY) as? Int {
             currentSetting = val
-            statusLabel.text = "Scheduled for \(minutesToTimeString(min: val))"
-        } else {
-            self.cancelButton.isEnabled = false
-            statusLabel.text = "Schedule a time:"
         }
         
         var components = DateComponents()
@@ -79,20 +66,8 @@ class ScheduleDailyDevotionViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
-    func configureBackButton() {
-        self.navigationItem.leftBarButtonItem = UIBarButtonItem(title: "Back", style: .done, target: self, action: #selector(self.backToDevotions(sender:)))
-    }
-    
-    @objc func backToDevotions(sender: UIBarButtonItem) {
-        MBStore.sharedStore.dispatch(PopCurrentNavigation())
-    }
-    
-    @IBAction func cancelNotifications(sender: UIBarButtonItem) {
-        center.removeAllPendingNotificationRequests()
-        // enter not-scheduled state
-        cancelButton.isEnabled = false
-        statusLabel.text = "Schedule a time:"
-        UserDefaults.standard.removeObject(forKey: MBConstants.DEFAULTS_DAILY_DEVOTION_TIME_KEY)
+    @IBAction func cancelPopover(sender: UIBarButtonItem) {
+        self.dismiss(animated: true, completion: nil)
     }
     
     @IBAction func scheduleNotifications(sender: UIBarButtonItem) {
@@ -100,14 +75,12 @@ class ScheduleDailyDevotionViewController: UIViewController {
         let components = NSCalendar.current.dateComponents(Set<Calendar.Component>([.hour, .minute]), from: self.timePicker.date)
         
         if let hour = components.hour, let minute = components.minute {
-            self.promptForNotifications(withDevotions: self.devotions, atHour: hour, minute: minute) { permission in
+            self.scheduler.promptForNotifications(withDevotions: self.devotions, atHour: hour, minute: minute) { permission in
                 DispatchQueue.main.async {
                     switch permission {
                     case .allowed:
-                        self.cancelButton.isEnabled = true
                         let totalMin = hour * 60 + minute
                         UserDefaults.standard.set(totalMin, forKey: MBConstants.DEFAULTS_DAILY_DEVOTION_TIME_KEY)
-                        self.statusLabel.text = "Scheduled for \(self.minutesToTimeString(min: totalMin))"
                     case .denied:
                         self.promptForSettings()
                     case .error:
@@ -134,11 +107,7 @@ class ScheduleDailyDevotionViewController: UIViewController {
                 }
                 
                 if UIApplication.shared.canOpenURL(settingsUrl) {
-                    if #available(iOS 10.0, *) {
-                        UIApplication.shared.open(settingsUrl)
-                    } else {
-                        UIApplication.shared.openURL(settingsUrl as URL)
-                    }
+                    UIApplication.shared.open(settingsUrl)
                 }
             }
         }))
@@ -153,58 +122,5 @@ class ScheduleDailyDevotionViewController: UIViewController {
         alert.addAction(UIAlertAction(title: "Dismiss", style: UIAlertActionStyle.default, handler: nil))
         self.present(alert, animated: true, completion: nil)
     }
-    
-    func promptForNotifications(withDevotions devotions: [LoadedDevotion], atHour hour: Int, minute: Int, completion: @escaping (NotificationPermission) -> Void) {
-        center.requestAuthorization(options: [.alert, .sound]) { granted, error in
-            if granted {
-                self.scheduleNotifications(withCenter: self.center, forDevotions: devotions, atHour: hour, minute: minute)
-                completion(.allowed)
-            } else if error != nil {
-                completion(.error)
-            } else {
-                completion(.denied)
-            }
-        }
-    }
-    
-    func scheduleNotifications(withCenter center: UNUserNotificationCenter, forDevotions devotions: [LoadedDevotion], atHour hour: Int, minute: Int) {
-        print("scheduling notifications")
-        
-        let startDate = Date().toMMddString()
-        let sortedDevotions = devotions.sorted {$0.date < $1.date}
-        guard let startIndex = (sortedDevotions.index { $0.dateAsMMdd == startDate }) else {
-            return
-        }
-        
-        var i = startIndex
-        var outstandingNotifications: Int = 0
-        
-        while outstandingNotifications < MBConstants.DEVOTION_NOTIFICATION_WINDOW_SIZE {
-            outstandingNotifications += 1
-            
-            let devotion = sortedDevotions[i]
-            let notificationId = "daily-devotion-\(devotion.dateAsMMdd)"
-            
-            if let dateComponents = devotion.dateComponentsForNotification(hour: hour, minute: minute) {
-                let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: false)
-                let content = DevotionNotificationContent(devotion: devotion)
-                let request = UNNotificationRequest(identifier: notificationId, content: content, trigger: trigger)
-                center.add(request)
-            }
-            
-            i = (i + 1) % sortedDevotions.count
-        }
-    }
-}
 
-extension Date {
-    static let ddMMFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "MM-dd"
-        return formatter
-    }()
-    
-    func toMMddString() -> String {
-        return Date.ddMMFormatter.string(from: self)
-    }
 }
